Adding final ensemble model
I have tried many combinations for the DNN structure but the one with ReLU activation works like a charm. But since the data points are meager once the final CSV is populated we can explore other configurations as well.
PS Create a Models folder and copy the models inside it.